requirements.txt

Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Werkzeug==2.2.3
PyMySQL==1.0.3
cryptography==40.0.2
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Pillow==9.5.0
itsdangerous==2.1.2
pyjwt==2.6.0